[build][test] guard native archive cache invalidation
